task3 first try create structure
