+++ Task3/src/app.ts
@@ -1,17 +1,8 @@
-import express, { Router, Application } from 'express';
-import { getUserById, getAutoSuggestUsers, addUser, editUser, deleteUser, isUserExist } from './appServices';
+import express, { Application } from 'express';
+import { UserController } from './controllers/controller';
 
-const app: Application = express();
-const router: Router = express.Router();
+export const app: Application = express();
 
 app.listen(3000, () => console.log('server running'));
 
-app.use(express.json());
-app.use('/', router);
-
-router.param('id', isUserExist);
-router.get('/user/:id', getUserById);
-router.get('/userSuggest', getAutoSuggestUsers);
-router.post('/user', addUser);
-router.put('/user/:id', editUser);
-router.delete('/user/:id', deleteUser);
+UserController();
