+++ Task3/src/types/types.ts
